feat: context -> zustand 로 변경
